requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
tqdm
wandb
matplotlib 
fiftyone
